Corrected Redraw
Apparently, Clear has problems when Transformation is active, because it
gets the device coords of the area and then clears the region defined by
these coords, interpreted as logical. This sounds odd, but it looks like
a bug in wxWidgets(?!). Anyway, the fix is to temporarily reset
transformation when doing Clear.