Graph algorithms + cleanup
Added a class to create and draw a boost graph and run several max flow
algorithms on it.
Added calls to exercise this functionality as unit tests in a new suite.
Added an example code from boost as another unit test for exploration.

Made changes to comply with the changes in the runner.